Add timestamp and data availability to pods
The existing code has been modified to include a timestamp in the pod state. This will be useful for tracking and debugging purposes. In addition, several changes were made to adjust key naming for data availability storage. Moreover, significant additions were implemented in the blocksync package for WASM block and transaction handling.

Co-authored-by: aakash4dev <aakash4dev.crypto@gmail.com>
Co-authored-by: Rahul Singh Maraskole <rsm050501@gmail.com>
Co-authored-by: Deadlium <uddeshjais.21@gmail.com>