Update configuration setup and command execution
This update enhances the configuration setup, command execution, as well as initiates changes in the Sequencer, P2P, and RPC implementations. A new TOML configuration file is added, commands are optimized to enhance execution, and several modifications have been done related to the sequencer operations. In addition, unused viper dependencies are replaced with go-toml. The 'init.go' and 'start.go' files are adjusted to reflect changes in configuration filepath.

Co-authored-by: noooblien <rsm050501@gmail.com>
Co-authored-by: aakash4dev <aakash4dev.crypto@gmail.com>
Co-authored-by: deadlium <uddeshjais.21@gmail.com>