Refactored logging levels and corrected initial transaction states
The current change set primarily includes the modification of many logging levels from 'Warn' to 'Debug' to reduce the verbosity of the application logs. It also addressed an error in the initial state of the transaction which was previously set to `InitVRF` and is now properly set to `PreInit`. Additional minor changes include commenting out unused code, deletion of unnecessary white space, and enhancement of logging details.