Update Makefile and command scripts to include build info
The Makefile was updated to gather additional build info such as commit date and branch name and to include this info in the build. During the build process, checks are now performed to ensure the working directory is clean before the build begins and the dependencies are up-to-date. The command scripts were adjusted to recognize this additional build information and output it when the version command is run.